+++ src/main/java/com/cognite/client/ApiBase.java
@@ -241,6 +241,15 @@ abstract class ApiBase {
             case EXTRACTION_PIPELINE:
                 itemReader = connector.readExtractionPipelinesById();
                 break;
+            case TRANSFORMATIONS:
+                itemReader = connector.readTransformationsById();
+                break;
+            case TRANSFORMATIONS_JOBS:
+                itemReader = connector.readTransformationJobsById();
+                break;
+            case TRANSFORMATIONS_SCHEDULES:
+                itemReader = connector.readTransformationSchedulesById();
+                break;
             default:
                 LOG.error(batchLogPrefix + "Not a supported resource type: " + resourceType);
                 throw new Exception(batchLogPrefix + "Not a supported resource type: " + resourceType);
@@ -436,6 +445,18 @@ abstract class ApiBase {
                 break;
             case THREED_ASSET_MAPPINGS:
                 results = connector.readThreeDAssetMappings(requestParameters);
+                break;
+            case TRANSFORMATIONS:
+                results = connector.readTransformations(requestParameters);
+                break;
+            case TRANSFORMATIONS_JOBS:
+                results = connector.readTransformationJobs(requestParameters);
+                break;
+            case TRANSFORMATIONS_JOB_METRICS:
+                results = connector.readTransformationJobMetrics(requestParameters);
+                break;
+            case TRANSFORMATIONS_SCHEDULES:
+                results = connector.readTransformationSchedules(requestParameters);
                 break;
             default:
                 throw new Exception("Not a supported resource type: " + resourceType);